Adds runbook for the alert PrometheusRuleFailures
Issue: https://issues.redhat.com/browse/OCPBUGS-4219

Problem: currently there is no clear direction on how to troubleshoot
the alert. This alert comes from the prometheus repo, so improving the
descripion there with OpenShift specific details doesn't seem to be the
best approach.

Solution: create a runbook to be linked in the OpenShift alert

Signed-off-by: JoaoBraveCoding <jmarcal@redhat.com>