alerts: ClusterVersionOperatorDown: Remove redundant section
The `oc get clusterversion` command is helpful for triaging
upgrade issues or to get further insights into how the CVO operates.

However, on its own, without further context, the command is
not necessarily helpful for triaging the ClusterVersionOperatorDown
alert, which signals that the CVO is down.

To keep the content of the runbook consistent while providing guidance
for cluster administrators, remove the brief section.