Remove travisn & leelavg as they are not member of openshift GH org
The following error is thrown when non openshift member are added

The following users are mentioned in OWNERS file(s) but are untrusted
for the following reasons. One way to make the user trusted is to add
them as members of the openshift org. You can then trigger verification
by writing /verify-owners in a comment.

travisn
User is not a member of the org. User is not a collaborator. Satisfy at
least one of these conditions to make the user trusted.
alerts/openshift-container-storage-operator/OWNERS
leelavg
User is not a member of the org. User is not a collaborator. Satisfy at
least one of these conditions to make the user trusted.
alerts/openshift-container-storage-operator/OWNERS

Signed-off-by: Malay Kumar Parida <mparida@redhat.com>